Fix viewing files in subdirs (invalid fix; do not use)
They were broken due to support for branches with slashes in the name,
which in routes that used branches were grabbing every slashed segment
up to the end.

By explicitly listing every known tag and branch in our regex, we
keep that from happening.

This approach is quite inelegant and does not work, causing crashes when
one repo has a branch named 'branch/with/slashes' and another contains a
branch named 'branch'.

Rather than matching a regex, the route should compute valid values for
the 'branch' param based on the 'repo' param.